pyproject.toml
[tool.poetry]
name = "prototrees"
version = "0.1.0"
description = ""
authors = ["DonQueso89 <kg.v.ekeren@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
protobuf = "^3.14.0"
grpcio = "^1.35.0"
grpcio-tools = "^1.35.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
